API Listing

WiNDCNational.National — Type

National

The primary container for national data tables. There are three fields, all dataframes:

data: The main data table.

sets: The sets table, describing the different sets used in the model.

elements: The elements table, describing the different elements in the model.

WiNDCNational.absorption_tax_rate — Method

absorption_tax_rate(data::AbstractNationalTable; column::Symbol = :value, output::Symbol = :value)

Calculate the absorption tax rate of the sectors.

Required Arguments

data::AbstractNationalTable: The national data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

Output

Returns a DataFrame with the columns :sectors and :value.

WiNDCNational.adjust_intermediate_flows — Method

adjust_intermediate_flows(X::National; kwargs...)

We take inputs to be negative and outputs to be positive. However, it is possible to have positive values in Intermediate_Demand and negative values in Intermediate_Supply. This function adjusts the flows accordingly.

Returns a National table.

WiNDCNational.armington_supply — Method

armington_supply(data::AbstractNationalTable; column::Symbol = :value, output::Symbol = :value)

Calculate the armington supply of the sectors.

Required Arguments

data::AbstractNationalTable: The national data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

Output

Returns a DataFrame with the columns :sectors and :value.

Process

\[\sum_{s\in\text{Sectors}} \text{Intermediate_Demand} + \text{Exogenous_Final_Demand} + \text{Personal_Consumption}\]

WiNDCNational.build_national_table — Method

build_national_table(aggregation::Symbol = :summary)
build_national_table(yaml_path::String)

Build and return the national table. By default this accepts a path to a YAML file describing the sets and parameters in the excel files. For convenience, you can also call this function with the :summary or :detailed symbols to load the corresponding YAML files.

For a full discussion on the structure of the YAML files, refer to the documentation.

WiNDCNational.create_margin_categories — Method

create_margin_categories(X::National; kwargs...)

Split both the Trade and Transport parameters in to Margin_Demand (positive values) and Margin_Supply (negative values).

Also create the set margin which points at the Transport and Trade sectors.

Return a National table.

WiNDCNational.create_pce_categories — Method

create_pce_categories(X::National; kwargs...)

Split Personal_Consumption parameters into Personal_Consumption (negative values) and Household_Supply (positive values).

Return a National table.

WiNDCNational.download_supply_use — Method

download_supply_use(use_pattern::Regex, supply_pattern::Regex; output_path::String = tempname())

Download the supply and use tables from the BEA. Return paths to each table.

Required Arguments

- use_pattern::Regex: A regex pattern to match the use tables.
- supply_pattern::Regex: A regex pattern to match the supply tables.

Optional Arguments

- output_path::String: The path to save the downloaded tables. Defaults to a temporary directory.

Return

A NamedTuple containing the use_path and supply_path.

WiNDCNational.fetch_supply_use — Method

fetch_supply_use(
    ;
    url::String = "https://apps.bea.gov/industry/iTables%20Static%20Files/AllTablesSUP.zip",
    output_path::String = tempname(),
)

Fetch the supply and use tables from the BEA website. The data is stored in a zip file, which is downloaded and extracted to the output_path. The extracted files are then returned as a vector of strings.

Optional Arguments

url::String: The url of the zip file containing the supply and use tables. Default is

"https://apps.bea.gov/industry/iTables%20Static%20Files/AllTablesSUP.zip".

output_path::String: The path to save the extracted files. Default is a temporary directory.

Output

Returns a vector of the absolute paths to the extracted files.

WiNDCNational.filter_to_sut — Method

filter_to_sut(x::AbstractString)

Return true if the string x corresponds to either a detailed or summary supply or use table. Otherwise, return false.

WiNDCNational.gross_output — Method

gross_output(data::AbstractNationalTable; column::Symbol = :value, output::Symbol = :value)

Calculate the gross output of the sectors.

Required Arguments

data::AbstractNationalTable: The national data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

Output

Returns a DataFrame with the columns :sectors and :value.

Process

Note that the below is negative as we take outputs to be negative and we want gross output to be positive.

\[-\sum_{s\in\text{Sectors}} \text{Intermediate_Supply} + \text{Household_Supply} - \text{Margin_Supply}\]

WiNDCNational.import_tariff_rate — Method

import_tariff_rate(data::AbstractNationalTable; column::Symbol = :value, output::Symbol = :value)

Calculate the import tariff rate of the sectors.

Required Arguments

data::AbstractNationalTable: The national data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

Output

Returns a DataFrame with the columns :sectors and :value.

WiNDCNational.load_national_yaml — Method

load_national_yaml(yaml_path::String)

Load the YAML file that describes the national tables. This function also ensures the file has the correct structure and fields.

Examples of this file are available for the summary data and detailed data.

Structure of YAML file

metadata

years: A list of years covered by the data (e.g., [2020, 2021, 2022]). Defaults to "all".

supply_path: The path to the supply data. Defaults to ""

use_path: The path to the use data. Defaults to ""

supply_pattern: A regex pattern to match the file name of the supply table. Defaults to ""

use_pattern: A regex pattern to match the file name of the use table. Defaults to ""

transformation_keywords: A list of keywords for data transformation. Defaults to []

You must either specify both use_path and supply_path or supply_pattern and use_pattern.

The transformation_keywords are used to pass parameters to the data transformation functions. In particular, it's useful to have an insurance_codes keyword for the redistribution of CIF_FOB data to imports and transport.

sets

Each set is listed with the following attributes:

description - A description for the set

domain - Which column the set operates on

values - An Excel range that corresponds to the NAICS codes

descriptions - An Excel range that corresponds to the descriptions of the NAICS codes

table - Either "use" or "supply"

parameters

Each parameter name is listed with the following attributes:

param - The name of the parameter. This acts as the set to access the parameter.

description - A description of the parameter

row - The row set of the parameter.

col - The column set of the parameter

table - Either "use" or "supply"

flipsign - Boolean, changes the sign of the data. Only used on `sectorsubsidy` since the values are reported as positive, but must be negative.

The parameters reference the sets, you shouldn't need to update the parameters if you only change the sets.

composite_parameters

Each composite parameter is listed with the following attributes:

description - A description of the composite parameter

elements - A list of elements that make up the composite parameter. These should be parameter names

WiNDCNational.output_tax_rate — Method

output_tax_rate(data::AbstractNationalTable; column = :value, output = :value)

Calculate the other tax rate of the sectors.

Required Arguments

data::AbstractNationalTable: The national data.

Keyword Arguments

column::Symbol = :value: The column to be used for the calculation.

output::Symbol = :value: The name of the output column.

Output

Returns a DataFrame with the columns :sectors and :value.

WiNDCNational.redistribute_cif_fob — Method

redistribute_cif_fob(X::National; insurance_codes::Vector{String} = [], kwargs...)

Redistribute the CIFFOB parameter to Transport and Import, add all insurance commodities values to Import and non-insurance commodities to Transport. Remove CIFFOB parameter.

Returns a National table.